requirements.txt
# Kafka
kafka-python==2.0.2

# Data generation
faker==19.13.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Spark (install PySpark separately — see README)
pyspark==3.5.0

# Snowflake
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# dbt
dbt-core==1.7.4
dbt-snowflake==1.7.1

# Airflow (install separately — see README)
apache-airflow==2.8.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.17.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
great-expectations==0.18.8
